+++ packages/medusa/src/models/product-category.ts
@@ -25,6 +25,9 @@ export class ProductCategory extends BaseEntity {
   @Column()
   name: string
 
+  @Column({ nullable: false, default: '' })
+  description: string
+
   @Index({ unique: true })
   @Column({ nullable: false })
   handle: string