+++ packages/modules-sdk/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @medusajs/modules-sdk
 
+## 1.8.2
+
+### Patch Changes
+
+- Updated dependencies [[`4f58ddee0`](https://github.com/medusajs/medusa/commit/4f58ddee03509a4c46af160e5824cba80d4c950a)]:
+  - @medusajs/types@1.8.2
+
+## 1.8.1
+
+### Patch Changes
+
+- Updated dependencies [[`654a54622`](https://github.com/medusajs/medusa/commit/654a54622303139e7180538bd686630ad9a46cfd), [`abdb74d99`](https://github.com/medusajs/medusa/commit/abdb74d997f49f994bff49787a396179982843b0)]:
+  - @medusajs/types@1.8.1
+  - @medusajs/utils@1.8.1
+
 ## 1.8.0
 
 ### Minor Changes